% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2.R
\name{bayes_H2}
\alias{bayes_H2}
\title{Title}
\usage{
bayes_H2(x, newdata = NULL, v = NULL, summary = FALSE, cred = 0.95)
}
\arguments{
\item{x}{}

\item{newdata}{}

\item{v}{}

\item{summary}{}

\item{cred}{}
}
\value{

}
\description{
Title
}
