% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I2.R
\name{bayes_I2}
\alias{bayes_I2}
\title{Title}
\usage{
bayes_I2(x, newdata = NULL, v = NULL, summary = FALSE, cred = 0.95)
}
\arguments{
\item{x}{}

\item{newdata}{}

\item{summary}{}
}
\value{

}
\description{
Title
}
