% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blsmeta-package.R
\docType{package}
\name{blsmeta-package}
\alias{blsmeta}
\alias{blsmeta-package}
\title{blsmeta: Bayesian Location-Scale Meta-Analysis}
\description{
Fit Bayesian meta-analytic models, including fixed-effects, two-level, 
    and three-level random-effects models. Moderators can be 
    included for both the location and scale parameters.
}
\author{
\strong{Maintainer}: Donald Williams \email{drwwilliams@ucdavis.edu}

Authors:
\itemize{
  \item Josue Rodriguez \email{jerrodriguez@ucdavis.edu }
}

}
\keyword{internal}
