% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blsmeta.R
\name{blsmeta}
\alias{blsmeta}
\title{Title}
\usage{
blsmeta(
  yi,
  vi,
  sei,
  es_id,
  study_id,
  mods = ~1,
  mods_scale2 = ~1,
  mods_scale3 = ~1,
  prior = NULL,
  iter = 5000,
  warmup = 1000,
  chains = 4,
  log_linear = TRUE,
  data
)
}
\arguments{
\item{yi}{}

\item{vi}{}

\item{sei}{}

\item{es_id}{}

\item{study_id}{}

\item{mods}{}

\item{mods_scale2}{}

\item{mods_scale3}{}

\item{prior}{}

\item{iter}{}

\item{warmup}{}

\item{chains}{}

\item{log_linear}{}

\item{data}{}
}
\value{

}
\description{
Title
}
