% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.blsmeta}
\alias{coef.blsmeta}
\title{Title}
\usage{
\method{coef}{blsmeta}(object, summary = TRUE, cred = 0.95, ...)
}
\arguments{
\item{object}{}

\item{summary}{}

\item{cred}{}

\item{...}{}
}
\value{

}
\description{
Title
}
