% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint_blsmeta.R
\name{confint.blsmeta}
\alias{confint.blsmeta}
\title{Credible Intervals for Heterogeneity Statistics}
\usage{
\method{confint}{blsmeta}(object, cred = 0.95, digits = 3, ...)
}
\arguments{
\item{object}{An object of class \code{blsmeta}.}

\item{cred}{numeric. credible interval (defaults to \code{0.95}).}

\item{digits}{numeric. The desired number of digits for the summarized
estimates (defaults to \code{3}).}

\item{...}{Currently ignored.}
}
\value{
An object of class \code{confint}, including a data.frame
with the estimates.
}
\description{
Compute credible intervals for between-study variance,
between-study standard deviation, I2, and H2. This
function mimics that of \strong{metafor}, but for credible
and not confidence intervals.
}
\examples{
library(psymetadata)

fit_re <- blsmeta(yi = yi, vi = vi, 
                  es_id = es_id,
                  data = gnambs2020)

confint(fit_re)
}
