% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity_stats.R
\name{credint}
\alias{credint}
\title{S3 \code{credint} method}
\usage{
credint(object, ...)
}
\arguments{
\item{object}{An object of class \code{blsmeta}}

\item{...}{Currently ignored}
}
\description{
S3 \code{credint} method
}
