% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_agadullina2018}
\alias{dat_agadullina2018}
\title{Studies on Out-Group Entitativity and Prejudice}
\format{
A data frame with 85 rows and 9 variables:
\itemize{
\item \code{es_id}: effect size id
\item \code{study_id}: study id
\item \code{author}: study author
\item \code{pub_year}: year of publication
\item \code{n}: sample size
\item \code{design}: within or between subjects design
\item \code{ent_alpha}: entitativity score
\item \code{yi}: effect size (Fisher z)
\item \code{vi}: sampling variance (SE^2)
}
}
\usage{
data("dat_agadullina2018")
}
\description{
Results from 21 studies, including 85 effect sizes (Fisher Z), on the effect of
out-Group entitativity and prejudice \insertCite{agadullina2018people}{blsmeta}.
}
\details{
Further details can be found at \href{https://osf.io/8dw5y/}{https://osf.io/8dw5y/}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
