% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_coles2019}
\alias{dat_coles2019}
\title{Studies on the Facial Feedback Literature}
\format{
A dataset with 286 rows and 52 variables.
\describe{
\item{study_id}{Unique id for study}
\item{es_id}{Unique id for effect size}
\item{yi}{Effect size in Cohen's d}
\item{vi}{Variance of effect size (SE^2)}
\item{title}{Title of publication}
\item{year}{Year of publication}
\item{file_drawer}{Publicatin status}
\item{prop_women}{Proportion of study that identified as women}
\item{video_know}{Yes: Participants were told they were going to be recorded or the methodology stated that a video camera was placed within participant view.
No" Methodology stated that participants were unaware of video recording, that the video camera was hidden, or that there was not a video camera}
\item{stim}{Type of stimuli}
\item{proc}{Type of facial feedback manipulation}
\item{proc_aware}{Whether participants were aware of the facial feedback manipulation}
\item{w_v_b}{Whether the study used a between- or within-subjects design}
}
}
\source{
\href{https://osf.io/v8kxb/}{Open Science Framework}
}
\usage{
data(dat_coles2019)
}
\description{
To be filled in.. \insertNoCite{coles2019meta}{blsmeta}
}
\references{
\insertAllCited{}
}
\keyword{datasets}
