% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_denizAksayli2019}
\alias{dat_denizAksayli2019}
\title{Studies on the cognitive and academic benefits of Cogmed}
\format{
A dataset with 637 rows and 39 variables.
\describe{
\item{study_id}{}
\item{es_id}{}
\item{yi}{Effect size in Hedge's g}
\item{vi}{Variance (SE^2)}
\item{ni}{Sample size}
\item{author}{}
\item{transfer}{Transfer type: near or far}
\item{test}{Type of working memory test?}
\item{allocation}{Whether participants were randomly assigned}
\item{comparison}{Active or non-active: whether the CWMT groups was compared to another cogntively demannding activity}
\item{baseline}{Standardized mean diference corrected for upward bias between exp. and control at pre-test assesment}
\item{age_group}{Whether particpants were children (< 16 yrs), adults (17-55), or older adults (> 55)}
\item{age_mean_exp}{}
\item{age_mean_control}{}
\item{population}{Whether the participants were typical subjects not suffering from any clinical conditions}
}
}
\source{
\href{https://osf.io/jhavp/}{Open Science Framework}
}
\usage{
data(dat_denizAksayli2019)
}
\description{
\insertNoCite{aksayli2019cognitive}{blsmeta}
}
\references{
\insertAllCited{}
}
\keyword{datasets}
