% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_gnambs2020}
\alias{dat_gnambs2020}
\title{Studies on the Color Red and Cognitive Performance}
\format{
A data frame with 67 rows and 10 variables:
\itemize{
\item \code{es_id}: effect size id
\item \code{study_id}: study id
\item \code{author}: study author
\item \code{pub_year}: year of publication
\item \code{country}: country where experiment was conducted
\item \code{color}: control color
\item \code{n}: sample size
\item \code{design}: within or between subjects design
\item \code{yi}: effect size (standardized mean difference)
\item \code{vi}: sampling variance (SE^2)
}
}
\usage{
data("dat_gnambs2020")
}
\description{
Results from 22 studies, including 67 effect sizes (SMD), on the effect of
the color red on cognitive performance \insertCite{gnambs2020limited}{blsmeta}.
}
\details{
Further details can be found at \href{https://psyarxiv.com/a4qdv/}{https://psyarxiv.com/a4qdv/}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
