% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_maldonado2018}
\alias{dat_maldonado2018}
\title{Studies on Age Differences in Executive Functioning}
\format{
A data frame with 1268 rows and 13 variables:
\itemize{
\item \code{es_id}: effect size id
\item \code{study_id}: study id
\item \code{author}: study authors
\item \code{domain}: executive functioning domain
\item \code{n1}: sample size in younger group
\item \code{n2}: sample size in older group
\item \code{n_total}: total sample size (n1 + n2)
\item \code{mean_age1}: mean age of younger group
\item \code{mean_age2}: mean age of older group
\item \code{miyake}: framework put forward by Mijake and colleagues
\item \code{task}: cognitive task administered
\item \code{yi}: effect size (Hedge's g)
\item \code{vi}: sampling variance (SE^2)
}
}
\usage{
data("dat_maldonado2018")
}
\description{
Results from 431 studies, including 1268 effect sizes (Hedge's g), on age differences in
executive functioning \insertCite{maldonado2020age}{blsmeta}.
}
\details{
Further details can be found at \href{https://osf.io/bcywg/}{https://osf.io/bcywg/}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
