% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_manylabs2018}
\alias{dat_manylabs2018}
\title{Studies from the Many Labs 2 project.}
\format{
A dataset with 1,414 rows and 23 variables.
\describe{
\item{lab}{The lab which conducted the replication}
\item{es_id}{Unique id for each effect size}
\item{yi_r}{A numeric indicating  the observed effect size, expressed in \emph{r}}
\item{vi_r}{A numeric indicating the variance on the observed effect size, expressed in \emph{r}}
\item{yi_d}{A numeric indicating the observed effect size, expressed in Cohen's \emph{d}}
\item{vi_d}{A numeric indicating the variance on the observed effect size, expressed in Cohen's \emph{d}}
\item{ni}{A numeric indicating the total sample size for the observed effect size}
\item{country}{Country where the sample was collected}
\item{weird}{Dummy variable encoding whether a country was classified as WEIRD, \code{0 = non-WEIRD, 1 = WEIRD}}
\item{western}{Dummy variable encoding a team judgment whether country was considered "western"}
\item{educated}{Education score as measured by the Education Index}
\item{industrialized}{Industrialization score as measured in the 2016 Industrial Development Report}
\item{rich}{Dummy variable encoding whether a country is developed according to the 2014 World Economic Situation and Prospects Report, \code{0 = emerging or in transition, 1 = developed}}
\item{democratic}{The quality democracy in the corresponding country according to the 2015 Democracy Ranking Report. Higher scores indicate higher quality.}
\item{mean_weird_score}{The arithmetic mean of the \code{weird}, \code{western}, \code{educated}, \code{industrialized}, and \code{rich} variables}
\item{online}{Whether the study was replicated in a lab or online}
\item{analysis}{Unique id for replicated study}
}
}
\source{
\href{https://osf.io/ux3eh/}{Open Science Framework}
}
\usage{
data(dat_manylabs2018)
}
\description{
A subset of the data collected in the Many Labs 2 project which conducted replications of 28 classic and contemporary findings in psychology.
The study examined the extent to which variability in replication success can be attributed to the study sample.
}
\references{
Klein, R. A., et al. (2018). Many Labs 2: Investigating variation in replicability across samples and settings.
\emph{Advances in Methods and Practices in Psychological Science}, \emph{1}(4), 443-490.
(\href{https://www.psychologicalscience.org/publications/many-labs-2}{APS})
}
\keyword{datasets}
