% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_noble2018}
\alias{dat_noble2018}
\title{Studies on Shared Reading and Language Development}
\format{
A data frame with 316 rows and 13 variables:
\itemize{
\item \code{es_id}: effect size id
\item \code{study_id}: study id
\item \code{author}: study author
\item \code{measure}: measure used in the study
\item \code{age}: age of participants, grouped into categories.
\item \code{ses}: socio-economic status
\item \code{experimenter}: who administered the test (pa)
\item \code{duratio}: number of weeks
\item \code{dialogic_reading}: dialogic reading
\item \code{follow_up}: follow up
\item \code{n}: sample size
\item \code{yi}: effect size (Hedge's g)
\item \code{vi}: sampling variance (SE^2)
}
}
\usage{
data("dat_noble2018")
}
\description{
Results from 54 studies, including 316 effect sizes (Hedge's g), on the extent to which
shared reading impacts language development \insertCite{noble2019impact}{blsmeta}.
}
\details{
Further details can be found at \href{https://osf.io/34xyw/}{https://osf.io/34xyw/}
}
\references{
\insertAllCited{}
}
\keyword{datasets}
