% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_nuijten2020}
\alias{dat_nuijten2020}
\title{Meta-analytic data collected from}
\format{
A dataset with 2443 rows and 23 variables.

\describe{
\item{study_id}{Unique id for study}
\item{effect_id}{Unique id for effect size}
\item{authors}{A factor encoding an identifier for the primary study within a meta-analysis based on
the first author of the study or the sample used}
\item{year}{A numeric indicating the year in which the primary study was reported}
\item{yi}{A numeric indicating the original effect size converted to a Fishers z value}
\item{vi}{A numeric indicating the variance around the z value in \code{yi}}
\item{ni}{A numeric indicating the total sample size of the primary study}
\item{es}{A numeric of the effect size as indicated in \code{type_es}}
\item{se}{A numeric of the standard errors of the effect size}
\item{type_es}{A factor indicating the type of effect size extracted from the meta-analysis.
\itemize{
\item{1 = r transformed to Fishers z}
\item{2 = Hedge's g}
\item{3 = log odds ratio}
\item{4 = Cohen's d}
\item{5 = Hazard Ratio}
\item{6, 7, 8 = other}
}}
\item{type}{A factor indicating why type of IQ research was summarized by the meta-analysis.
\itemize{
\item{1 = Correlational}
\item{2 = Group_differences}
\item{3 = Experiments/Interventions}
\item{4 = Toxicology}
\item{5 = (Behavior) Genetics}
}}
\item{citations}{A numeric indicating number of times the primary study was cited}
\item{countrycode}{A factor the country in which the first author of a primary study was situated at the time of publication}
\item{jrnl_impact}{A numeric indicating the impact factor in 2014 of the journal where the primary study was published}
\item{similarity}{A factor indicating whether the primary study matched the research question of the meta-analysis
\itemize{\item{0 = dissimilar} \item{1 = similar}}}
}
}
\source{
\href{https://osf.io/fq5wp/}{Open Science Framework}
}
\usage{
data(dat_nuijten2020)
}
\description{
To be filled in...

\insertNoCite{nuijten2020effect}{blsmeta}
}
\references{
\insertAllCited{}
}
\keyword{datasets}
