% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_steffens2020}
\alias{dat_steffens2020}
\title{Studies on Social Identity Theory and Leadership: Leader Group Prototypicality}
\format{
A data frame with 251 rows and 10 variables:
\itemize{
\item \code{es_id}: effect size id
\item \code{study_id}: study id
\item \code{author}: study author
\item \code{n}: sample size
\item \code{design}: 0 = experimental; 1 = correlational
\item \code{published}: 0 = published; 1 = unpublished
\item \code{proto_strength}: 0 = ad-hoc; 1 = natural
\item \code{target.leader}: 0 = informal; 1 = formal
\item \code{yi}: effect size (Fisher z)
\item \code{vi}: sampling variance (SE^2)
}
}
\usage{
data("dat_steffens2020")
}
\description{
Results from 128 studies, including 251 effect sizes (Fisher Z), on the extent to which
a leader is perceived to embody shared social identity \insertCite{steffens2021advancing}{blsmeta}.
}
\details{
Further details can be found at \href{https://osf.io/hxgpw/}{https://osf.io/hxgpw/}
}
\references{
\insertAllCited{}
}
\keyword{datasets}
