% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_wibbelink2017}
\alias{dat_wibbelink2017}
\title{Studies on juvenile recidivism}
\format{
A dataset with 100 rows and 10 variables.
\describe{
\item{study_id}{unique id for each study}
\item{es_id}{unique id for each effect size}
\item{yi}{observed effect sizes(Cohen's \emph{d})}
\item{vi}{sampling variance (SE^2)}
\item{pstatpub}{dummy variable encoding whether the study was published, \code{0 = unpublished, 1 = published}}
\item{pstatnotpub}{dummy variable encoding whether the study was unpublished, \code{0 = published, 1 = unpublished}}
\item{typgen}{dummy variable encoding the type of recidivism behavior \code{0 = not applicable, 1 = general}}
\item{typovert}{dummy variable encoding the type of recidivism behavior \code{0 = not applicable, 1 = overt}}
\item{typcovert}{dummy variable encoding the type of recidivism behavior \code{0 = not applicable, 1 = covert}}
\item{pyear}{the publication year of the study; mean-centered}
}
}
\source{
\href{https://www.tqmp.org/RegularArticles/vol12-3/p154/}{The Quantitative Methods in Psychology}
}
\usage{
data(dat_wibbelink2017)
}
\description{
Meta-analytic data on studies investigating associations between mental health disorders of
delinquent juveniles and subsequent delinquent behavior
of those juveniles
}
\references{
Wibbelink et al. (2017). A meta-analysis of the association between mental health disorders and juvenile recidivism.
\emph{Aggression and Violent Behavior}, \emph{33}, 78-90.
(\href{https://www.sciencedirect.com/science/article/pii/S1359178917300149}{ScienceDirect})
}
\keyword{datasets}
