% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_blsmeta.R
\name{fitted.blsmeta}
\alias{fitted.blsmeta}
\title{Expected Values of the Predictive Distribution}
\usage{
\method{fitted}{blsmeta}(
  object,
  newdata = NULL,
  es_id = TRUE,
  study_id = TRUE,
  summary = TRUE,
  cred = 0.95,
  digits = 3,
  ...
)
}
\arguments{
\item{object}{An object of class \code{blsmeta}.}

\item{newdata}{An optional data.frame for which to compute
fit fitted values. Defaults to \code{NULL}, which then
uses the original data used in \code{\link[blsmeta]{blsmeta}}.}

\item{es_id}{logical. Should the level two "random" effect be included
(defaults to \code{TRUE})}

\item{study_id}{logical. Should the level three "random" effects be included
(defaults to \code{TRUE})?}

\item{summary}{logical. Should the posterior samples be summarized
(defaults to \code{TRUE})?}

\item{cred}{numeric. credible interval (defaults to \code{0.95}).}

\item{digits}{numeric. The desired number of digits for the summarized
estimates (defaults to \code{3}).}

\item{...}{Currently ignored.}
}
\value{
A data frame of fitted values (when \code{summary = TRUE}),
}
\description{
Compute the expected (analogous to "fitted") values
for a \code{blsmeta} object.
}
\examples{
library(psymetadata)

dat <- gnambs2020

fit <- blsmeta(yi = yi,
               vi = vi,
               es_id = es_id,
               data = dat)
               
fitted(fit)
}
