% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model.R
\name{get_model}
\alias{get_model}
\title{Get JAGS Code from Fitted Model}
\usage{
get_model(object)
}
\arguments{
\item{object}{An object of class \code{blsmeta}}
}
\value{
JAGS code printed out.
}
\description{
Get JAGS Code from Fitted Model
}
\examples{
library(psymetadata)

fit <-  blsmeta(yi = yi, 
                vi = vi, 
                mods = ~n,
                es_id = es_id,
                data = gnambs2020)
                
get_model(object)
}
