% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prior.R
\name{get_prior}
\alias{get_prior}
\title{Get Prior from Fitted Model}
\usage{
get_prior(object)
}
\arguments{
\item{object}{An object of class \code{blsmeta}.}
}
\value{
model priors are printed out.
}
\description{
This function returns the prior distributions, and
can be used to check that the priors are correctly specified
(e.g., when supplying custom priors).
}
\examples{
library(psymetadata)

fit <-  blsmeta(yi = yi, 
                vi = vi, 
                mods = ~n,
                es_id = es_id,
                data = gnambs2020)
                
get_prior(object)
}
