% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2.R
\name{H2}
\alias{H2}
\title{H2 Heterogeneity Statistic}
\usage{
H2(
  object,
  newdata_scale2 = NULL,
  newdata_scale3 = NULL,
  s2 = NULL,
  cred = 0.95,
  summary = TRUE,
  digits = 3
)
}
\arguments{
\item{object}{An object of class \code{blsmeta}.}

\item{newdata_scale2}{An optional data.frame for which to compute
predictions for the level 2 variance component.
Defaults to \code{NULL}, which then uses the
original data used in \code{\link[blsmeta]{blsmeta}}}

\item{newdata_scale3}{An optional data.frame for which to compute
predictions for the level 3 variance component.
Defaults to \code{NULL}, which then uses the
original data used in \code{\link[blsmeta]{blsmeta}}}

\item{s2}{numeric. A user-defined "typical" sampling variance.
Defaults to the estimator in Equation 9 of
\insertCite{Higgins2002a;textual}{blsmeta}.}

\item{cred}{numeric. credible interval (defaults to \code{0.95}).}

\item{summary}{logical. Should the posterior samples be summarized
(defaults to \code{TRUE})?}

\item{digits}{numeric. The desired number of digits for the summarized
estimates (defaults to \code{3}).}
}
\value{
A data frame of predicted values.
}
\description{
Compute H2 (total variability divided by sampling variability).
Note that the H2 formulation for three-level models was inspired by
\insertCite{cheung2014modeling;textual}{blsmeta}.
}
\note{
The sampling variances are assumed to be known.
}
\examples{
# data
library(psymetadata)

# no scale model
fit <- blsmeta(yi, vi,
               es_id = es_id,
               data = gnambs2020)

# compute I2 for all data
h2 <- H2(fit)

# scale model
fit <- blsmeta(yi, vi,
               es_id = es_id,
               mods_scale2 = ~ n,
               data = gnambs2020)

new_data <- data.frame(n = c(100, 150))

# compute I2 for new data
h2 <- H2(fit, newdata_scale2 = new_data)
}
\references{
\insertAllCited{}
'
}
