% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc.R
\name{ICC}
\alias{ICC}
\title{Intraclass Correlation Coefficients}
\usage{
ICC(
  object,
  newdata_scale2 = NULL,
  newdata_scale3 = NULL,
  cred = 0.95,
  summary = TRUE,
  percent = TRUE,
  digits = 3
)
}
\arguments{
\item{object}{An object of class \code{blsmeta}.}

\item{newdata_scale2}{An optional data.frame for which to compute
predictions for the level 2 variance component.
Defaults to \code{NULL}, which then uses the
original data used in \code{\link[blsmeta]{blsmeta}}.}

\item{newdata_scale3}{An optional data.frame for which to compute
predictions for the level 3 variance component.
Defaults to \code{NULL}, which then uses the
original data used in \code{\link[blsmeta]{blsmeta}}}

\item{cred}{numeric. credible interval (defaults to \code{0.95}).}

\item{summary}{logical. Should the posterior samples be summarized
(defaults to \code{TRUE})?}

\item{percent}{logical. Should the results be percentages, as in \strong{metafor}
(defaults to \code{TRUE})?}

\item{digits}{numeric. The desired number of digits for the summarized
estimates (defaults to \code{3}).}
}
\value{
A list of two data frames of predicted values.
}
\description{
Compute ICCs for three-level meta-analysis
Note that the ICC formulation for three-level models is described in
\insertCite{cheung2014modeling;textual}{blsmeta}. It is defined as
the heterogeneity at each level divided by the total heterogeneity,
which will approach I2 when the
sampling variances approach zero (large n).
}
\details{
In essence, with a scale model, this results in the
ICCs being a function of moderators. Further, rather than one
ICC, there is an ICC for each of the \code{k} studies.
For more information about varying ICCs,
we refer interested users to \insertCite{williams2019putting;textual}{blsmeta}
and \insertCite{williams2020fine}{blsmeta}.
}
\examples{
library(psymetadata)

fit <- blsmeta(yi = yi, vi = vi, 
               es_id = es_id,
               study_id = study_id,
               data = gnambs2020)


iccs <- ICC(object = fit, 
            summary = TRUE)
}
\references{
\insertAllCited{}
}
