% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_hypothesis.R
\name{linear_hypothesis}
\alias{linear_hypothesis}
\title{Test Linear Hypotheses for Posterior Samples}
\usage{
linear_hypothesis(
  lin_comb,
  obj,
  cred = 0.95,
  sub_model = NULL,
  print_names = FALSE,
  rope = NULL,
  contrast = NULL
)
}
\arguments{
\item{lin_comb}{A string specifying a linear combination of variables,
or a list of variable names if using \code{contrast}.}

\item{obj}{An object of class \code{blsmeta}}

\item{cred}{The level for which a credible interval should be computed.}

\item{sub_model}{character. Which sub-model for the hypotheses, with options
including \code{location} and \code{scale}. The defaul
is \code{NULL}, so either is required.}

\item{print_names}{logical. Should the parameter names be printed ?
This is useful for specifying the hypotheses,
as the names need to be exact.}

\item{rope}{Specify a ROPE. Optional.}

\item{contrast}{A contrast matrix specifying which combinations to test. Optional.}
}
\value{
An object of class \code{linear_hypothesis}
}
\description{
Test Linear Hypotheses for Posterior Samples
}
\examples{
library(psymetadata)

fit <-  blsmeta(yi = yi, 
                vi = vi, 
                es_id = es_id,
                study_id = study_id,
                data = gnambs2020)
                
linear_hypothesis(obj = fit, 
                  lin_comb = "scale3_Intercept > scale2_Intercept",
                  sub_model = "scale")
}
