% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prior.R
\name{make_prior}
\alias{make_prior}
\title{Make Prior Distributions}
\usage{
make_prior(
  yi,
  vi,
  sei,
  es_id,
  study_id,
  mods = ~1,
  mods_scale2 = ~1,
  mods_scale3 = ~1,
  prior = NULL,
  iter = 5000,
  warmup = 1000,
  chains = 4,
  log_linear = TRUE,
  data
)
}
\arguments{
\item{yi}{vector with the observed effect sizes (length \code{k}).}

\item{vi}{vector with the sampling variances (length \code{k}).
Provide either \code{vi} or \code{sei}.}

\item{sei}{vector with the sampling variances (length \code{k}).
Provide either \code{vi} or \code{sei}.}

\item{es_id}{numeric vector with the effect size ids (i.e., \code{1:k}). When
provided, a two-level random-effects model is estimated.
Whereas, when not specified, a fixed-effects model is estimated
by default.}

\item{study_id}{numeric vector with the study ids (length \code{k}). When provided,
a three-level random-effects model is estimated
(\code{es_id} is required). Note that the \code{yi}'s are
assumed to be nested within \code{study_id}.}

\item{mods}{an object of class \code{\link[stats]{formula}},
including moderator(s) for the effect size. By default, an
intercept only model is fitted (i.e., \code{mods = ~ 1}),
resulting in the overall effect.}

\item{mods_scale2}{an object of class \code{\link[stats]{formula}},
including moderator(s) for level two variance component
(or "scale"). By default, an intercept only model is fitted
(i.e., \code{mods = ~ 1}), resulting in the customary estimate
(assumed constant across studies).}

\item{mods_scale3}{an object of class \code{\link[stats]{formula}},
including moderator(s) for level three variance component
(or "scale"). By default, an intercept only model is fitted
(i.e., \code{mods = ~ 1}), resulting in the customary estimate
(assumed constant across studies). See \strong{Details}.}

\item{prior}{one or more \code{blsmetaprior} objects created by
\code{set_prior}.}

\item{iter}{numeric. The number of posterior samples per chain
(defaults to \code{5000}, excluding \code{warmup}).}

\item{warmup}{numeric. The number of warmup samples, which are discarded
(defaults to \code{1000}).}

\item{chains}{numeric. The number of chains (defaults to \code{4})}

\item{log_linear}{logical. Should the variance components be modeled on
the log-scale (defaults to \code{TRUE})? This is
applicable to the scale models with no moderators.
See \strong{Details}.}

\item{data}{data frame containing the variables in the model.}
}
\value{
A character string with the priors.
}
\description{
Make user-defined prior distributions. This
is useful for ensuring the priors are correctly specified,
and also sampling from the priors.
}
\examples{
library(psymetadata)

prior <- 
  c(assign_prior(param = "(Intercept)", 
                 prior = "dnorm(0, 1)", dpar = "location"),
                 assign_prior(param = "(Intercept)", 
                 prior = "dnorm(0, 10)", 
                 dpar = "scale", level = "two")
                 )

make_prior(yi = yi, 
           vi = vi, 
           prior = prior,
           es_id = es_id,
           study_id = study_id,
           data = gnambs2020)
}
