% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_study_id.R
\name{make_study_id}
\alias{make_study_id}
\title{Make \code{study_id} Variable}
\usage{
make_study_id(x)
}
\arguments{
\item{x}{character vector}
}
\value{
numeric \code{study_id}
}
\description{
\link[blsmeta]{blsmeta} requires a numeric vector
for the \code{study_id} variable create that variable from
a character vector (for example).
}
\examples{
# data
library(psymetadata)

make_study_id( gnambs2020$author )
}
