% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_rma.R
\name{mcmc_rma}
\alias{mcmc_rma}
\title{MCMC Sampling for \code{rma} Objects}
\usage{
mcmc_rma(object, data, ...)
}
\arguments{
\item{object}{an object of class \code{rma}.}

\item{data}{data frame containing the variables in the model.}

\item{...}{Currently ignored}
}
\value{
An object of class \code{blsmeta} (see \code{\link[blsmeta]{blsmeta}})
}
\description{
The package \strong{metafor} is perhaps the gold-standard for
meta-analysis in \code{R}. This function allows for sampling
from the posterior distribution of a model originally estimated
with \strong{metafor}.
}
\note{
This function works for any kind of model fitted with \code{rma}.
}
\examples{
library(metafor)
library(psymetadata)

fit <- mcmc_rma(rma(yi = yi, vi = vi, 
                method = "FE",
                data = gnambs2020), 
                data = gnambs2020)

fit
}
