% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predict.R
\name{posterior_predict}
\alias{posterior_predict}
\title{Title}
\usage{
posterior_predict(
  object,
  mod = NULL,
  mod_tau_2 = NULL,
  data = NULL,
  re_formula = NULL,
  summary = TRUE,
  cred = 0.95,
  ...
)
}
\arguments{
\item{object}{}

\item{mod}{}

\item{mod_tau_2}{}

\item{data}{}

\item{re_formula}{}

\item{summary}{}

\item{cred}{}

\item{...}{}
}
\value{

}
\description{
Title
}
