% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_samples.R
\name{posterior_samples}
\alias{posterior_samples}
\title{Extract Posterior Samples}
\usage{
posterior_samples(object, param = NULL, print_names = FALSE)
}
\arguments{
\item{object}{An object of class \code{blsmeta}.}

\item{param}{character string. The parameter(s) to be extracted.
See \strong{Details}}

\item{print_names}{logical. Should the parameter names be printed
(defaults to \code{FALSE})? This is useful
for seeing which parameters can be extracted.}
}
\value{
A \code{iter} * \code{chains}  by parameter matrix of posterior
samples
}
\description{
Extract posterior samples from an
object of class \code{blsmeta}
}
\details{
Options for \code{param} depend on the model, but can included
printed out by setting \code{print_names = TRUE}.

Note the variance components, and log-linear coefficients are on
the standard deviation scale.
}
\examples{
library(psymetadata)

fit <-  blsmeta(yi = yi, 
                vi = vi, 
                es_id = es_id,
                data = gnambs2020)
                
posterior_samples(object = fit, 
                  print_names = TRUE)
}
