% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.blsmeta}
\alias{predict.blsmeta}
\title{Title}
\usage{
\method{predict}{blsmeta}(x, mod, mod_tau_2, data)
}
\arguments{
\item{x}{}

\item{mod}{}

\item{mod_tau_2}{}

\item{data}{}
}
\value{

}
\description{
Title
}
