% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_interval.R
\name{predict}
\alias{predict}
\title{Posterior Predictive Distribution}
\usage{
predict(
  object,
  newdata = NULL,
  newdata_scale2 = NULL,
  newdata_scale3 = NULL,
  summary = TRUE,
  new_level = TRUE,
  cred = 0.95,
  digits = 3,
  ...
)
}
\arguments{
\item{object}{An object of class \code{blsmeta}.}

\item{newdata}{An optional data.frame for which to compute
fit fitted values. Defaults to \code{NULL}, which then
uses the original data used in \code{\link[blsmeta]{blsmeta}}.}

\item{newdata_scale2}{An optional data.frame for which to compute
predictions for the level 2 variance component.
Defaults to \code{NULL}, which then uses the
original data used in \code{\link[blsmeta]{blsmeta}}.}

\item{newdata_scale3}{An optional data.frame for which to compute
predictions for the level 3 variance component.
Defaults to \code{NULL}, which then uses the
original data used in \code{\link[blsmeta]{blsmeta}}}

\item{summary}{logical. Should the posterior samples be summarized
(defaults to \code{TRUE})?}

\item{new_level}{logical. Should the predictive distribution be for
new levels (defaults to \code{TRUE})? See \strong{Details}.}

\item{cred}{numeric. credible interval (defaults to \code{0.95}).}

\item{digits}{numeric. The desired number of digits for the summarized
estimates (defaults to \code{3}).}

\item{...}{Currently ignored.}
}
\value{
A data frame of fitted values (when \code{summary = TRUE}), or
a matrix of the unsummarized posterior samples.
}
\description{
Draw sample from the posterior predictive distribution.
}
\details{
When \code{new_level = TRUE}, this provides the predictive
distribution with respect to the moderators for the sub-models.
For obtaining the predictive distribution for the observed
levels, set this to \code{FALSE} which then includes the
"random" effects.
}
\examples{
library(psymetadata)

fit <-  blsmeta(yi = yi, 
                vi = vi, 
                es_id = es_id,
                data = gnambs2020)
                
predict(fit)
}
