% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_blsmeta.R
\name{print.blsmeta}
\alias{print.blsmeta}
\title{Print \code{blsmeta} Objects}
\usage{
\method{print}{blsmeta}(x, ...)
}
\arguments{
\item{x}{An object of class \code{blsmeta}}

\item{...}{Ignored}
}
\description{
Print \code{blsmeta} Objects
}
