% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_hypothesis.R
\name{print.linear_hypothesis}
\alias{print.linear_hypothesis}
\title{Print formatted summary of a \code{linear_hypothesis} object}
\usage{
\method{print}{linear_hypothesis}(x, ...)
}
\arguments{
\item{x}{An object of class \code{linear_hypothesis}}

\item{...}{Other arguments to be passed to \code{print}}
}
\value{
A formatted summary of posterior samples
}
\description{
Print formatted summary of a \code{linear_hypothesis} object
}
