% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psuedo_r2.R
\name{pseudo_R2}
\alias{pseudo_R2}
\title{Title}
\usage{
pseudo_R2(
  fit_re,
  fit_me,
  mod_tau_2 = NULL,
  summary = TRUE,
  cred = 0.95,
  data,
  measure = "r2"
)
}
\arguments{
\item{fit_re}{}

\item{fit_me}{}

\item{mod_tau_2}{}

\item{summary}{}

\item{data}{}

\item{measure}{}
}
\value{

}
\description{
Title
}
