% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psuedo_r2.R
\name{psuedo_R2}
\alias{psuedo_R2}
\title{Title}
\usage{
psuedo_R2(fit_re, fit_me, mod_tau_2 = NULL, data, measure = "r2")
}
\arguments{
\item{fit_re}{}

\item{fit_me}{}

\item{mod_tau_2}{}

\item{data}{}

\item{measure}{}
}
\value{

}
\description{
Title
}
