% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranef.R
\name{ranef.blsmeta}
\alias{ranef.blsmeta}
\title{Title}
\usage{
\method{ranef}{blsmeta}(object, summary = TRUE, cred = 0.95, ...)
}
\arguments{
\item{object}{}

\item{summary}{}

\item{cred}{}

\item{...}{}
}
\value{

}
\description{
Title
}
