% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranef_blsmeta.R
\name{ranef}
\alias{ranef}
\alias{ranef.blsmeta}
\title{Extract Group-Level Estimates}
\usage{
\method{ranef}{blsmeta}(object, summary = TRUE, cred = 0.95, digits = 3, ...)
}
\arguments{
\item{object}{An object of class \code{blsmeta}}

\item{summary}{logical. Should the posterior samples be summarized
(defaults to \code{TRUE})?}

\item{cred}{numeric. credible interval (defaults to \code{0.95}).}

\item{digits}{numeric. The desired number of digits for the summarized
estimates (defaults to \code{3}).}

\item{...}{Currently ignored.}
}
\value{
A list of class \code{blsmeta}. When \code{summary = TRUE},
each object is a data frame including the summarized posterior
samples. When \code{summary = FALSE}, each object is a matrix
with the posterior samples.
}
\description{
Extract the group-level ("random") effects from
a \code{blsmeta} object.
}
\examples{
library(psymetadata)
dat <- gnambs2020

fit <- blsmeta(yi = yi, 
               vi = vi, 
               es_id = es_id, 
               data = dat)

ranef(fit)
}
