% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_prior.R
\name{sample_prior}
\alias{sample_prior}
\title{Draw Samples from the Prior Distributions}
\usage{
sample_prior(object, iter = 5000, ...)
}
\arguments{
\item{object}{An object of class \code{make_prior}.}

\item{iter}{numeric. The number of posterior samples per chain
(defaults to \code{5000}, excluding \code{warmup}).}

\item{...}{Currently ignored.}
}
\value{
A data frame including the posterior samples.
}
\description{
This function is useful for visualizing the
prior distributions.
}
\details{
Note that parameters include
\code{beta} (location or effect size),
\code{gamma} (level-two scale model),
and \code{eta} (level-three scale model). Importantly,
the scale model priors are on the log-scale (use \code{exp}).
}
\examples{
library(psymetadata)

prior <- c(assign_prior(param = "(Intercept)", 
           prior = "dnorm(0, 1)", dpar = "location"),
           assign_prior(param = "(Intercept)", 
           prior = "dnorm(-2, 1)", 
           dpar = "scale", level = "two")
           )

priors <- make_prior(yi = yi, 
                     vi = vi, 
                     prior = prior,
                     es_id = es_id,
                     study_id = study_id,
                     data = gnambs2020)

samps <- sample_prior(priors, iter = 50000)
hist(samps$beta)
}
