% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_prior.R
\name{set_prior}
\alias{set_prior}
\title{Title}
\usage{
set_prior(param, prior, dpar, level = NULL)
}
\arguments{
\item{param}{}

\item{prior}{}

\item{dpar}{}

\item{level}{}
}
\value{

}
\description{
Title
}
