% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkage_factor.R
\name{shrinkage_factor}
\alias{shrinkage_factor}
\title{Title}
\usage{
shrinkage_factor(x, mod_tau2 = NULL, summary = TRUE, cred = 0.95, data)
}
\arguments{
\item{x}{}

\item{mod_tau2}{}

\item{summary}{}

\item{cred}{}

\item{data}{}
}
\value{

}
\description{
Title
}
