% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau.R
\name{tau}
\alias{tau}
\title{Title}
\usage{
tau(x, mod_tau_2 = NULL, data, summary = TRUE, cred = 0.95)
}
\arguments{
\item{x}{}

\item{mod_tau_2}{}

\item{data}{}

\item{summary}{}

\item{cred}{}
}
\value{

}
\description{
Title
}
