% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau2.R
\name{tau2}
\alias{tau2}
\title{Variance Components}
\usage{
tau2(
  object,
  type = "sd",
  newdata_scale2 = NULL,
  newdata_scale3 = NULL,
  cred = 0.95,
  summary = TRUE,
  digits = 3
)
}
\arguments{
\item{object}{An object of class \code{blsmeta}.}

\item{type}{character. Should the variance or standard deviation? The options
are \code{type = "var"} and \code{type = "sd"} (the default).}

\item{newdata_scale2}{An optional data.frame for which to compute
predictions for the level 2 variance component.
Defaults to \code{NULL}, which then uses the
original data used in \code{\link[blsmeta]{blsmeta}}}

\item{newdata_scale3}{An optional data.frame for which to compute
predictions for the level 3 variance component.
Defaults to \code{NULL}, which then uses the
original data used in \code{\link[blsmeta]{blsmeta}}}

\item{cred}{numeric. credible interval (defaults to \code{0.95}).}

\item{summary}{logical. Should the posterior samples be summarized
(defaults to \code{TRUE})?}

\item{digits}{numeric. The desired number of digits for the summarized
estimates (defaults to \code{3}).}
}
\value{
A data frame of predicted values.
}
\description{
Typically, the variance components are assumed constant
across the \code{k} studies in meta-analysis. When scale modeling,
the variance components are a function of
moderators, which can be computed with this function
\insertCite{@see @williams2021putting}{blsmeta}.
}
\examples{

library(psymetadata)

fit <- blsmeta(yi = yi, vi = vi, 
               es_id = es_id,
               mods_scale2 = ~n,
               data = gnambs2020)


tau2(object = fit, 
     newdata_scale2 = data.frame(n = seq(20, 100, 10)))
}
\references{
\insertAllCited{}
'
}
