% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_variance.R
\name{total_variability}
\alias{total_variability}
\title{Title}
\usage{
total_variability(x, measure = "tau2", summary = TRUE, cred = 0.95)
}
\arguments{
\item{x}{}

\item{summary}{}

\item{cred}{}
}
\value{

}
\description{
Title
}
