% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_yi.R
\name{variance_yi}
\alias{variance_yi}
\title{Compute the Total Variance in \code{yi}}
\usage{
variance_yi(object, type = "sd", cred = 0.95, summary = TRUE)
}
\arguments{
\item{object}{object of class \code{blsmeta}.}

\item{type}{character. Should the variance or standard deviation? The options
are \code{type = "var"} and \code{type = "sd"} (the default).}

\item{cred}{numeric. credible interval (defaults to \code{0.95}).}

\item{summary}{logical. Should the posterior samples be summarized
(defaults to \code{TRUE})?}
}
\value{
Either a summarized data frame, including the
posterior mean, sd, and credible intervals, or a matrix
of dimensions \code{iter} * \code{chains} by \code{k}.
}
\description{
Typically, the variance components are assumed constant
across the \code{k} studies in meta-analysis, implying that
the total variance for each effect size is the sum of the
sampling and between-study variance. When scale modeling,
the variance in \code{yi} is now a function of those
moderators, which can be computed with this function
\insertCite{@see @williams2021putting}{blsmeta}.
}
\note{
The sampling variances are assumed to be known.
}
\examples{
library(psymetadata)

# no scale model
fit <- blsmeta(yi, vi, 
               es_id = es_id, 
               mods_scale2 = ~ 1,
               data = gnambs2020)

no_scale_mod <- variance_yi(fit)

# scale model
fit <- blsmeta(yi, vi, 
               es_id = es_id, 
               mods_scale2 = ~ n,
               data = gnambs2020)

scale_mod <- variance_yi(fit)
}
\references{
\insertAllCited{}
}
