% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blsmeta.R
\name{blsmeta}
\alias{blsmeta}
\title{Title}
\usage{
blsmeta(
  yi,
  vi,
  es_id = NULL,
  mod,
  scale_level_two,
  iter = 5000,
  chains = 4,
  data
)
}
\arguments{
\item{yi}{}

\item{vi}{}

\item{es_id}{}

\item{mod}{}

\item{scale_level_two}{}

\item{iter}{}

\item{chains}{}

\item{data}{}
}
\value{

}
\description{
Title
}
