% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blsmeta.R
\name{blsmeta}
\alias{blsmeta}
\title{Bayesian Meta-Analysis via (Mixed-Effects) Location-Scale Models}
\usage{
blsmeta(
  yi,
  vi,
  sei,
  es_id,
  study_id,
  mods = ~1,
  mods_scale2 = ~1,
  mods_scale3 = ~1,
  prior = NULL,
  iter = 5000,
  warmup = 1000,
  chains = 4,
  log_linear = TRUE,
  data
)
}
\arguments{
\item{yi}{vector with the observed effect sizes (length \code{k}).}

\item{vi}{vector with the sampling variances (length \code{k}).
Provide either \code{vi} or \code{sei}.}

\item{sei}{vector with the sampling variances (length \code{k}).
Provide either \code{vi} or \code{sei}.}

\item{es_id}{numeric vector with the effect size ids (i.e., \code{1:k}). When
provided, a two-level random-effects model is estimated.
Whereas, when not specified, a fixed-effects model is estimated
by default.}

\item{study_id}{numeric vector with the study ids (length \code{k}). When provided,
a three-level random-effects model is estimated
(\code{es_id} is required). Note that the \code{yi}'s are
assumed to be nested within \code{study_id}.}

\item{mods}{an object of class \code{\link[stats]{formula}},
including moderator(s) for the effect size. By default, an
intercept only model is fitted (i.e., \code{mods = ~ 1}),
resulting in the overall effect.}

\item{mods_scale2}{an object of class \code{\link[stats]{formula}},
including moderator(s) for level two variance component
(or "scale"). By default, an intercept only model is fitted
(i.e., \code{mods = ~ 1}), resulting in the customary estimate
(assumed constant across studies).}

\item{mods_scale3}{an object of class \code{\link[stats]{formula}},
including moderator(s) for level three variance component
(or "scale"). By default, an intercept only model is fitted
(i.e., \code{mods = ~ 1}), resulting in the customary estimate
(assumed constant across studies). See \strong{Details}.}

\item{prior}{one or more \code{blsmetaprior} objects created by
\code{\link[blsmeta]{assign_prior}}.}

\item{iter}{numeric. The number of posterior samples per chain
(defaults to \code{5000}, excluding \code{warmup}).}

\item{warmup}{numeric. The number of warmup samples, which are discarded
(defaults to \code{1000}).}

\item{chains}{numeric. The number of chains (defaults to \code{4})}

\item{log_linear}{logical. Should the variance components be modeled on
the log-scale (defaults to \code{TRUE})? This is
applicable to the scale models with no moderators.
See \strong{Details}.}

\item{data}{data frame containing the variables in the model.}
}
\value{
An object of class \code{blsmeta}. This is used internally,
and it is not all that useful otherwise.
}
\description{
Fit meta-analytic models, including fixed-effects, two-level,
and three-level random-effects models. Moderators can be
included for both the location and scale parameters. This
is accomplished with mixed-effects location-scale modeling
\insertCite{@see for example @Hedeker2008}{blsmeta}, with the basic
idea extended to meta-analysis in
\insertCite{williams2021putting;textual}{blsmeta}.
}
\details{
"\strong{Scale}"

The scale corresponds to the variance of a normal distribution. However,
in \strong{blsmeta} it is modeled on the standard deviation scale. As a
result, the reported estimates are also on the standard deviation
(log) scale. To make sense of the estimates, it is helpful to
use the predict function.

\strong{log_linear}

In the two and three-level models, by default a log-linear model is fitted
to the random-effects variances ("scale"). When no moderators are included in
\code{mods_scale2} and \code{mods_scale3}, this is an intercept only model
(the "scale" is constant across the \code{k} studies).

To use a different prior distribution (as opposed to the default log-normal),
this can be changed by setting \code{log_linear = FALSE}. In this case,
a half Student-t prior is employed which is then similar to the R
package \strong{brms}. Note that a log-linear model is required when
moderators are included in \code{mods_scale2} and \code{mods_scale3}.

\strong{mods_scale3}

Moderators for \code{mods_scale3} are inherently level 3 predictors.
This means that the study-level characteristic cannot \strong{not} vary within study.
For example, with, say, 100 effect sizes from 25 studies, this variance component is
predicted with only the 25 studies. To do so, the first row of every study
is used by default.
}
\note{
Three-level meta-analyses are described in \insertCite{van2013three;textual}{blsmeta},
\insertCite{cheung2014modeling;textual}{blsmeta}, and
\insertCite{assink2016fitting;textual}{blsmeta}.They allow for modeling dependent
effect sizes (several from the same study).
}
\examples{
# data
library(psymetadata)

  
fit <- blsmeta(yi = yi, vi = vi, 
               es_id = es_id, 
               data = gnambs2020, 
               chains = 2)
               
}
\references{
\insertAllCited{}
}
