% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_blsmeta.R
\name{coef.blsmeta}
\alias{coef.blsmeta}
\title{Extract Group-Specific Coefficients}
\usage{
\method{coef}{blsmeta}(object, summary = TRUE, cred = 0.95, digits = 3, ...)
}
\arguments{
\item{object}{An object of class \code{blsmeta}.}

\item{summary}{logical. Should the posterior samples be summarized
(defaults to \code{TRUE})?}

\item{cred}{numeric. credible interval (defaults to \code{0.95}).}

\item{digits}{numeric. The desired number of digits for the summarized
estimates (defaults to \code{3}).}

\item{...}{Currently ignored.}
}
\value{
A list of class \code{blsmeta}. When \code{summary = TRUE},
each object is a data frame including the summarized posterior
samples. When \code{summary = FALSE}, each object is a matrix
with the posterior samples.
}
\description{
Extract the group-specific coefficients, defined as
the sum of the fixed effects and corresponding random effects.
}
\examples{
...
}
