% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_gamble2019}
\alias{dat_gamble2019}
\title{Meta-analytic data collected from}
\format{
\describe{
\item{study_id}{Unique id for study}
\item{samp_id}{Unique id for each sample}
\item{es_id}{Unique id for effect size}
\item{authors}{Authors of study}
\item{yi}{Effect size in r}
\item{vi}{Variance of effect size}
\item{ni}{Sample size of study}
\item{sex}{Proportion of study that was female}
\item{age}{Mean age of participants}
\item{dep_status}{Clinical status of depression}
\item{comorbid_anx}{Whehter comorbid with anxiety}
\item{emo_val}{Emotional valence of simulations}
\item{macro_micro}{Macro vs. micro specificity}
\item{cue_type}{Cue type}
\item{spec_rated}{Self- vs. researcher-rated specificity}
\item{dep_rated}{Self- vs. researcher-rated depression}
\item{cat_dim}{Categorical vs. dimensional designs}
\item{quality}{Study quality rating}
\item{published}{Published or not}
\item{mode}{Mode or prospection}
}
}
\source{
\href{https://osf.io/5wjb2/}{Open Science Framework}
}
\usage{
data(dat_gamble2019)
}
\description{
To be filled in.. \insertNoCite{gamble2019specificity}{blsmeta}
}
\references{
\insertAllCited{}
}
\keyword{datasets}
