% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.blsmeta}
\alias{fitted.blsmeta}
\title{Title}
\usage{
\method{fitted}{blsmeta}(object, re_formula = NULL, summary = TRUE, cred = 0.95, ...)
}
\arguments{
\item{object}{}

\item{re_formula}{}

\item{summary}{}

\item{cred}{}

\item{...}{}
}
\value{

}
\description{
Title
}
