% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_blsmeta.R
\name{print.blsmeta}
\alias{print.blsmeta}
\title{Print \code{blsmeta} Objects}
\usage{
\method{print}{blsmeta}(x, cred = 0.95, ...)
}
\arguments{
\item{x}{An object of class \code{blsmeta}}

\item{cred}{numeric. credible interval (defaults to \code{0.95}).}

\item{...}{currently ignored.}
}
\description{
Print \code{blsmeta} Objects
}
